feat: replace EmptyBodyResponse into (), update readme.md for usage
