+++ Cargo.toml
@@ -12,7 +12,7 @@ repository = "https://github.com/gongzhengyang/axum-restful"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-aide = { version = "0.12", features = ["redoc", "macros", "axum-extra-query"] }
+aide = { version = "0.12", features = ["redoc", "macros", "axum-extra-query", "axum"] }
 anyhow = "1.0"
 async-trait = "0.1"
 axum = "0.6"
