+++ examples/demo/.env
@@ -1,7 +1,1 @@
-# config the base pg connect params
-POSTGRES_DB=demo
-POSTGRES_USER=demo-user
-POSTGRES_PASSWORD=demo-password
-
-# used by axum-restful framework to specific a database connection
-DATABASE_URL=postgres://${POSTGRES_USER}:${POSTGRES_PASSWORD}@localhost:5432/${POSTGRES_DB}
+DATABASE_URL=postgres://demo-user:demo-password@localhost:5432/demo
